backend/pyproject.toml
[project]
name = "s2s-gemini-live"
version = "0.1.0"
description = "Speech-to-Speech application using Gemini Live API and WebSocket"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "google-auth>=2.27.0",
    "google-cloud-aiplatform>=1.38.0",
    "python-dotenv>=1.0.0",
    "websockets>=12.0",
    "pydantic>=2.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
